refactor: ディレクトリ構造の整理とモジュール分割
- src/featuresのコンポーネントをsrc/componentsに移動
- 型定義とサービスをsrc/modulesに集約
- publicディレクトリをsrc配下に移動
- wxt.configの設定を整理（srcDir指定、不要な設定削除、vite用のコンフィグが無いとアイコンアクションがなくなる問題をMEMO）
- 不要なworker.tsを削除"
