+++ src/entrypoints/sidepanel/index.html
