+++ src/entrypoints/sidepanel/main.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import App from '../../src/components/App'
+import App from '../../components/App'
 import './style.css'
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
