+++ src/entrypoints/sidepanel/style.css
