+++ src/public/icon/96.png
